[refactor] s3 위치 리팩토링(#48)
