--- src/main/java/org/recordy/server/external/exception/ExternalException.java
@@ -1,11 +1,0 @@
-package org.recordy.server.external.exception;
-
-import org.recordy.server.common.exception.RecordyException;
-import org.recordy.server.common.message.ErrorMessage;
-
-public class ExternalException extends RecordyException {
-
-    public ExternalException(ErrorMessage errorMessage) {
-        super(errorMessage);
-    }
-}
