--- src/main/java/org/recordy/server/keyword/repository/KeywordRepository.java
@@ -1,11 +1,0 @@
-package org.recordy.server.keyword.repository;
-
-import org.recordy.server.keyword.domain.Keyword;
-
-import java.util.List;
-
-public interface KeywordRepository {
-
-    // query
-    List<Keyword> findAll();
-}
