--- src/main/java/org/recordy/server/keyword/service/KeywordService.java
@@ -1,11 +1,0 @@
-package org.recordy.server.keyword.service;
-
-import org.recordy.server.keyword.domain.Keyword;
-
-import java.util.List;
-
-public interface KeywordService {
-
-    // query
-    List<Keyword> getAll();
-}
