+++ src/main/java/org/recordy/server/record/repository/impl/RecordQueryDslRepository.java
@@ -3,14 +3,13 @@ package org.recordy.server.record.repository.impl;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import lombok.RequiredArgsConstructor;
 import org.recordy.server.common.util.QueryDslUtils;
-import org.recordy.server.keyword.domain.KeywordEntity;
-import org.recordy.server.keyword.domain.QKeywordEntity;
+import org.recordy.server.record.service.keyword.domain.KeywordEntity;
 import org.recordy.server.record.domain.QRecordEntity;
 import org.recordy.server.record.domain.QUploadEntity;
 import org.recordy.server.record.domain.RecordEntity;
+import org.recordy.server.record.service.keyword.domain.QKeywordEntity;
 import org.recordy.server.user.domain.QUserEntity;
 import org.recordy.server.user.domain.UserEntity;
-import org.recordy.server.user.repository.UserRepository;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
 import org.springframework.data.domain.SliceImpl;
