+++ src/main/java/org/recordy/server/record/service/RecordService.java
@@ -1,6 +1,6 @@
 package org.recordy.server.record.service;
 
-import org.recordy.server.keyword.domain.Keyword;
+import org.recordy.server.record.service.keyword.domain.Keyword;
 import org.recordy.server.record.domain.File;
 import org.recordy.server.record.domain.Record;
 import org.recordy.server.record.domain.usecase.RecordCreate;
