+++ src/main/java/org/recordy/server/record/service/impl/RecordServiceImpl.java
@@ -3,7 +3,7 @@ package org.recordy.server.record.service.impl;
 import lombok.RequiredArgsConstructor;
 import org.recordy.server.common.exception.RecordyException;
 import org.recordy.server.common.message.ErrorMessage;
-import org.recordy.server.keyword.domain.Keyword;
+import org.recordy.server.record.service.keyword.domain.Keyword;
 import org.recordy.server.record.domain.File;
 import org.recordy.server.record.domain.Record;
 import org.recordy.server.record.domain.usecase.RecordCreate;
