+++ src/main/java/org/recordy/server/record/service/keyword/repository/KeywordRepository.java
@@ -1,0 +1,11 @@
+package org.recordy.server.record.service.keyword.repository;
+
+import org.recordy.server.record.service.keyword.domain.Keyword;
+
+import java.util.List;
+
+public interface KeywordRepository {
+
+    // query
+    List<Keyword> findAll();
+}
