+++ src/main/java/org/recordy/server/record/service/keyword/repository/impl/KeywordJpaRepository.java
@@ -1,6 +1,6 @@
-package org.recordy.server.keyword.repository.impl;
+package org.recordy.server.record.service.keyword.repository.impl;
 
-import org.recordy.server.keyword.domain.KeywordEntity;
+import org.recordy.server.record.service.keyword.domain.KeywordEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface KeywordJpaRepository extends JpaRepository<KeywordEntity, Long> {
