+++ src/main/java/org/recordy/server/record/service/keyword/repository/impl/KeywordRepositoryImpl.java
@@ -1,9 +1,9 @@
-package org.recordy.server.keyword.repository.impl;
+package org.recordy.server.record.service.keyword.repository.impl;
 
 import lombok.RequiredArgsConstructor;
-import org.recordy.server.keyword.domain.Keyword;
-import org.recordy.server.keyword.domain.KeywordEntity;
-import org.recordy.server.keyword.repository.KeywordRepository;
+import org.recordy.server.record.service.keyword.domain.Keyword;
+import org.recordy.server.record.service.keyword.domain.KeywordEntity;
+import org.recordy.server.record.service.keyword.repository.KeywordRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
