+++ src/main/java/org/recordy/server/record/service/keyword/service/KeywordService.java
@@ -1,0 +1,11 @@
+package org.recordy.server.record.service.keyword.service;
+
+import org.recordy.server.record.service.keyword.domain.Keyword;
+
+import java.util.List;
+
+public interface KeywordService {
+
+    // query
+    List<Keyword> getAll();
+}
