+++ src/main/java/org/recordy/server/record/service/keyword/service/impl/KeywordServiceImpl.java
@@ -1,9 +1,9 @@
-package org.recordy.server.keyword.service.impl;
+package org.recordy.server.record.service.keyword.service.impl;
 
 import lombok.RequiredArgsConstructor;
-import org.recordy.server.keyword.domain.Keyword;
-import org.recordy.server.keyword.repository.KeywordRepository;
-import org.recordy.server.keyword.service.KeywordService;
+import org.recordy.server.record.service.keyword.domain.Keyword;
+import org.recordy.server.record.service.keyword.repository.KeywordRepository;
+import org.recordy.server.record.service.keyword.service.KeywordService;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
