+++ src/main/java/org/recordy/server/record_stat/repository/BookmarkQueryDslRepository.java
@@ -4,10 +4,7 @@ import com.querydsl.jpa.impl.JPAQueryFactory;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.recordy.server.common.util.QueryDslUtils;
-import org.recordy.server.keyword.domain.QKeywordEntity;
 import org.recordy.server.record.domain.QRecordEntity;
-import org.recordy.server.record.domain.QUploadEntity;
-import org.recordy.server.record.domain.RecordEntity;
 import org.recordy.server.record_stat.domain.BookmarkEntity;
 import org.recordy.server.record_stat.domain.QBookmarkEntity;
 import org.recordy.server.user.domain.QUserEntity;
