+++ src/test/java/org/recordy/server/mock/FakeContainer.java
@@ -13,9 +13,9 @@ import org.recordy.server.auth.service.impl.token.AuthTokenParser;
 import org.recordy.server.auth.service.impl.token.AuthTokenServiceImpl;
 import org.recordy.server.auth.service.impl.token.AuthTokenSigningKeyProvider;
 import org.recordy.server.external.service.S3Service;
-import org.recordy.server.keyword.repository.KeywordRepository;
-import org.recordy.server.keyword.service.KeywordService;
-import org.recordy.server.keyword.service.impl.KeywordServiceImpl;
+import org.recordy.server.record.service.keyword.repository.KeywordRepository;
+import org.recordy.server.record.service.keyword.service.KeywordService;
+import org.recordy.server.record.service.keyword.service.impl.KeywordServiceImpl;
 import org.recordy.server.mock.auth.FakeAuthApplePlatformServiceImpl;
 import org.recordy.server.mock.auth.FakeAuthKakaoPlatformServiceImpl;
 import org.recordy.server.mock.auth.FakeAuthRepository;
