+++ src/test/java/org/recordy/server/mock/keyword/FakeKeywordRepository.java
@@ -1,7 +1,7 @@
 package org.recordy.server.mock.keyword;
 
-import org.recordy.server.keyword.domain.Keyword;
-import org.recordy.server.keyword.repository.KeywordRepository;
+import org.recordy.server.record.service.keyword.domain.Keyword;
+import org.recordy.server.record.service.keyword.repository.KeywordRepository;
 
 import java.util.List;
