+++ src/test/java/org/recordy/server/record/domain/UploadEntityTest.java
@@ -1,8 +1,8 @@
 package org.recordy.server.record.domain;
 
 import org.junit.jupiter.api.Test;
-import org.recordy.server.keyword.domain.Keyword;
-import org.recordy.server.keyword.domain.KeywordEntity;
+import org.recordy.server.record.service.keyword.domain.Keyword;
+import org.recordy.server.record.service.keyword.domain.KeywordEntity;
 import org.recordy.server.util.DomainFixture;
 
 import static org.assertj.core.api.Assertions.assertThat;
